refactor: global components folder
