+++ src/app/(frontend)/(protected)/profile/page.tsx
@@ -1,5 +1,8 @@
 import ClientPage from "./client-page";
 
+// requires user to be logged in
+// requires user to have a profile
+// see (protected)/profile/layout.tsx
 export default function ProfilePage() {
     return <ClientPage />;
 }
