+++ src/app/(frontend)/(staff)/staff/memberCSVUpload/components/CSVReader.tsx
@@ -3,9 +3,9 @@
 import React, { useState } from "react";
 import { PaymentMethod } from "@libs/types/invoice.type";
 import { ProfileDTO } from "@libs/types/profile.type";
-import Button from "@ui/button/Button";
 import Papa from "papaparse";
 
+import Button from "@/components/button/Button";
 import { validateProfile } from "@/services/profile/validateProfile";
 
 import { generateMemberProfile } from "../services/tempGenerateMemberProfile";
