+++ src/app/(frontend)/not-found.tsx
@@ -1,14 +1,14 @@
-import Button from "@ui/button/Button";
+import Button from "@/components/button/Button";
 
 export default function NotFound() {
-  return (
-    <div className="flex items-center justify-center min-h-screen bg-black text-white text-lg">
-      <div className="flex flex-col items-center justify-center gap-10">
-        <h3>404 - Page Not Found</h3>
-        <Button href="/" variant={{ style: "solid" }}>
-          Return Home
-        </Button>
-      </div>
-    </div>
-  );
+    return (
+        <div className="flex min-h-screen items-center justify-center bg-black text-lg text-white">
+            <div className="flex flex-col items-center justify-center gap-10">
+                <h3>404 - Page Not Found</h3>
+                <Button href="/" variant={{ style: "solid" }}>
+                    Return Home
+                </Button>
+            </div>
+        </div>
+    );
 }
