+++ src/components/assets/image/generate-index.js
