+++ src/components/assets/image/index.ts
