+++ src/components/button/SignOut.tsx
@@ -1,7 +1,8 @@
 "use client";
 
 import { authClient } from "@libs/auth/auth-client";
-import Button from "@ui/button/Button";
+
+import Button from "@/components/button/Button";
 
 export default function SignOut() {
     return (
