+++ src/components/email/README.md
@@ -1,0 +1,5 @@
+These set of components are for emailing purpose
+See [page](http://localhost:3000/test) for email previews
+
+all are created using 
+[react.email](https://react.email/)
