+++ src/components/text/SlideInText.tsx
